Support reading levels from Mario's Picross for Game Boy.
This plugin supports only the US version of the game, though it might be trivial
to implement support for the JP version. The massage_out for writing the levels
back to the game is in place, but there's no function to actually do it, yet. It
will need to be tested, first.