Now MariosPicross can re-insert puzzles. Puzzle format has changed.
Previously, puzzles had a series of rows stored in the main dict. Now, they are
stored as a list of lists under the single key 'puzzle'. The work of getting the
puzzles into this format is handled by read_puzzles and write_puzzles.

Documentation was updated to match the new format, and the tests were updated as
well.